refactor(cms): colocate recipe field defs with Recipes collection
Move recipeIngredient/instruction Payload fields out of shared fields/ so
only the cross-collection slug helper remains there.
